+++ tests/Fixtures/WidgetService.php
@@ -2,9 +2,9 @@
 
 namespace Ngos\AdminCore\Tests\Fixtures;
 
-use Ngos\AdminCore\Services\CrudService;
+use Ngos\AdminCore\Services\BaseService;
 
-class WidgetService extends CrudService
+class WidgetService extends BaseService
 {
     public function __construct(Widget $widget)
     {